Reorganized and formatted the blockchain-albatross crate.
